pyproject.toml
[project]
name = "m4-2-advanced-middleware-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arctrl>=2.5.1",
    "cryptography>=45.0.6",
    "fastapi>=0.116.1",
    "python-gitlab>=6.2.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
]

